requirements.txt
discord.py==2.7.1
asyncpg==0.31.0
boto3==1.42.83
python-dotenv==1.2.2
aiohttp==3.13.5